wip youmu a bomb

+++ src/player.h
@@ -149,7 +149,9 @@ void player_add_bombs(Player *plr, int bombs);
 void player_add_points(Player *plr, unsigned int points);
 
 void player_cancel_bomb(Player *plr, int delay);
-int player_get_bomb_progress(Player *plr, double *out_speed);
+
+// Progress is normalized from 0: bomb start to 1: bomb end
+double player_get_bomb_progress(Player *plr, double *out_speed);
 int player_run_bomb_logic(Player *plr, void *ent, complex *argptr, int (*callback)(void *ent, int t, double speed));
 
 void player_preload(void);
